+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - translated variation configurator 
 - translated short and technical description 
+- removed comment and address rest call
 
 ## [4.6.0]
 ### Fixed